+++ src/OpenTelemetry.Exporter.InMemory/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+* Adds the ability to configure `MetricReaderOptions` via the
+  `AddInMemoryExporter` extension method.
+  ([#2931](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2931))
+
 ## 1.2.0-rc2
 
 Released 2022-Feb-02